requirements.txt
ollama
termcolor
google.generativeai
python-dotenv
python-telegram-bot